+++ src/adjust/adjusthandler.cpp
