+++ src/adjust/adjustproxyconnection.cpp
