+++ src/adjust/adjustproxyconnection.h
